stock/requirements.txt
futu-api==3.12.0
future==0.17.1
numpy==1.17.0
pandas==0.25.0
pync==2.0.3
